Update pon 13 11

+++ cms-new/admin/includes/view_all_users.php
@@ -82,11 +82,16 @@ if (isset($_GET['change_to_sub'])) {
 
 // delete user   
 if (isset($_GET['delete'])) {
-    $the_user_id = $_GET['delete'];
-    
-    $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
-    $delete_user_query = mysqli_query($connection, $query); 
-    header("Location: users.php"); // mora refreshati stranu da se vidi bez deletanog
+  if (isset($_SESSION['user_role'])) {
+    if ($_SESSION['user_role'] == 'admin') {  // da ne deletaju iz url-a sa sql injection
+
+        $the_user_id = mysqli_real_escape_string($connection, $_GET['delete']);
+        
+        $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
+        $delete_user_query = mysqli_query($connection, $query); 
+        header("Location: users.php"); // mora refreshati stranu da se vidi bez deletanog
+    } 
+}
 }
